requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-cryptography==1.1
django-dotenv==1.4.2
django-jazzmin==2.5.0
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
sqlparse==0.4.2
tzdata==2022.1
